fix(cache) do not set success types for records we're not resolving
If the additional section would contain a TXT record, then that would
be set as the success type, and `last` would try and look for that.
The cache lookup would succeed, but since it doesn't have a `address`
field, `toip()` would return `nil`.

fixes https://github.com/Kong/kong/issues/3210