chore(logging) add substantial debug logging
Especially in the async query part this is necessary to track
the exact order of events.
Since the messages extensivly use json serialization they are all
in long-comment brackets, which should ease to disable them
in released code. To be manually turned on by the user by doing a
"replace all" on the client code.