Fixes a bug for replacing stale data in the cache (#42)
records from the additional section would not be stored in the
cache when there already was a cache entry. This also means that
stale data was not replaced. The bug surfaced when a dns server
would not respond to CNAME queries, but only provide that data
alongside an A query.

To track the bug extensive logging was added (to be enabled manually),
and a script to re-format those logs.