fix(balancer) introduce a handle to set peer status
Setting status wasn't always possible, if an address still
contained a name instead of an ip (SRV -> A -> ip).
By returning a handle, the proper reference is preserved, and
as a side effect the `retryCount` parameter is now obsolete.

Signature of `getPeer` changed, and hence this is a breaking
change