feat(health) prevent cascading failures
When the balancer reaches an unhealthy state, it will stop
returning peers and error out on each request until it returns
to healthy again.

Note: this introduces a new error message that may be returned;
  balancer.error.ERR_BALANCER_UNHEALTHY = "Balancer is unhealthy"

Additional in this commit is some test refactoring such that
generic tests are now executed against all different types of
balancers/algorithms.