fix(balancer) use proper dns client
The balancer was calling `toip` on the dns-client module, instead
of using the dns client passed to it.
Since the module is to be used as a singleton anyway, this did not
affect any real usage. But in testing the cache would not be
properly reset, since there we reload the entire module forcefully.