fix(base) reduce the amount of timers used
The current implementation creates one recurring timer per balancer.

This is ok in many circumstances, but can exhaust nginx's allocated
resources.

This PR changes that by a single timer which updates all balancers
sequentially. When there's a need for updating a host, *that* part
gets delegated to an ephemeral (non-recurring) timer.

This way only 1 recurring timer gets created, and ephemeral
timers are created only when hosts need to be updated, and then
released.