fix(base) reduce the amount of timers used
This PR changes the timers so that instead of creating one per
balancer we crete a single global one. The price payed is that
we must store a `glocal` variable with all the lbs inside.

The DNS refresh is async so no further timers/threads are needed.

Notice that the single global recurring timer is now marked as
"detached" so that it is not garbage-collected.